Don't auto-retry RpcChannel_Send when error is final.
RpcChannel_Send retries in case of any error. However, retry makes
sense only when there is a transport error. When there is a failure
returned from VMware in handling the RPC command, the error is final
and retrying does not make any sense. We need to avoid retry when
it is not a transport error. In order to differentiate between
the two types of errors the internal function signatures required
changes. As there are not many direct consumers for RpcOut_send, fix
all of its callers to check rpcStatus along with the return value of
the function. New API behavior also allows removal of some string
comparisons on "reply".

There are no other behavioral changes intended in RpcChannel_Send,
so all the callers of this function remain untouched.