Remove an assert and save vSock family reference in vmsvc only.
VMCISock_GetAFValueFd API does not open a vsock device if kernel
has upstreamed drivers. So, the ASSERT for vsock fd is not valid
in such cases and has to go. Instead, save the reference only when
family as well as fd are valid, because there is no need to save
the reference in case of upstreamed drivers case anyway.

If 'vmusr' holds a reference to vSock address family, it could
come in the way of upgrade because 'vmusr' keeps running during
upgrade. To avoid that issue and given that we just need one
reference, we limit the reference to main service only.

Also fix the VMCISock_GetAFValueFd API to initialize outFd in
case of upstreamed drivers.