Increase the scope of 'opLock' to avoid the race in VmBackupAsyncCallback.
A previous change fixed a race between worker thread and main thread
VmBackupAsyncCallback that could lead to premature release of
'currentOp' from 'gBackupState'. If the 'currentOp' was set after
VmBackupAsyncCallback had queried status of the 'currentOp' but
VmBackupAsyncCallback was not done with 'status' value check,
it could release the 'currentOp' prematurely because 'status'
was initialized to VMBACKUP_STATUS_FINISHED. The fix was to
initialize the 'status' to VMBACKUP_STATUS_PENDING. The fix
broke tests that invoke quiescing with 'execScripts = false'
option and in that case there is no 'currentOp' setup by the
VmBackupStartScripts() call that leaves VmBackupAsyncCallback
in its initialized 'status' VMBACKUP_STATUS_PENDING forever.
This fix reverts the previous fix by restoring the 'status' initial
value to VMBACKUP_STATUS_FINISHED and increases the scope of the
lock 'opLock' to address the race condition.

This change also adds some Tools configuration settings to make
it easy to test HBR scenario.

As reading configuration is common functionality, moved
the common functions to vmtoolslib.