Fix HGFS mount call to create mount point
For FUSE clients and open-vm-tools installations the package
might not create the shared folders mount point. The only reliable
way to ensure that we have one is to create it ourselves.
This seems okay to do since it is a hardcoded path and has always
been set to this same path without any regard for configuration by
users, although that may change in later change sets.

This also ensures if a user has inadvertently deleted the mount
point after installing tools and before enabling the Shared Folders
feature the mount will be robust enough to still mount.

Secondly, did some clean up of the mount function to move the check
for already mounted into a separate function. This will prove useful
if a umount call is added on a disable of the feature.

Finally, the mount file table check now uses the common mount info
macros and the system calls are replaced with ProcMgr exec calls.
The file access checks now use the glib routines as do the mkdir and
rmdir for mount point creation and deletion.