Make Vsock RPC code handle the ECONNRESET
VMX might delay releasing a vsock when the client side closes
the end point first.  This makes a later client vsock connect
attempt fail with ECONNRESET using the same source port as the
closed one.  Note that the client binds OK on the reused source
port, since the client side has released the socket on that port.

Rewrote the connect code as to clarify it. Explicitly separated
the retry logic.  Added code to return the system errno, and
added additional information about which socket API call failed.