Hgfs Fuse Client: fix attribute caching from hardcoded 5 seconds
The default attribute caching of 5 seconds is too long and is not
cleared by modifications to the file. This leaves users' applications
receiving stale attribute information for a file which can cause
applications to crash.

Set the default setting for the cache timeout to 1 so that corresponds
to the same as the kernel client.

Fix the rename that purges the cache of stale attribute data for the
target of the rename.