Hgfs Win Client: fix upgrade of driver by changing the Hgfs install settings
Some upgrades of the vmhgfs driver are requiring an additional
reboot even after the initial reboot. The Difx APIs replace the
driver with the new one on reboot, but after the old one is loaded.
This means that after the reboot everything looks good on the surface
but the old driver is running and not even on the disk at that point.
To mitigate this we can change the driver start type and fix missed
install settings both of which makes the HGFS driver compatible with
the rest of the Windows redirectors.

Perform to system registry changes to the vmhgfs driver service:
- set the HGFS driver dependent on the MUP component which it is anyway.
- set the driver start type to be demand start and not a standard system driver.

This makes that the vmtools system service responsible for starting the
Shared Folders client driver. This is again more compatible with the
Windows redirector model.