Tools RpcIn: fix an assert triggered by a race condition
RpcIn open channel asserts that the current channel is NULL. This was being
triggered under stress testing.  The guest tools service ran slowly, so
the VMX timed out and reset the channel. This caused the tools service
to do a channel check reset and which then called the RpcChannelRestart to
stop and start the channel again. Unfortunately, this call modified the
channel global without synchronizing with the other threads sending RPCs.
DiskWipe plugin was in the middle of a RpcChannel_Send (with the lock
acquired) when the restart call ran concurrently.

To fix this issue the RpcChannelRestart now acquires the same channel lock
to around the channel stop and start calls so that all sending threads wait
or complete first.