Tools RPC: Fix XDR invocation on mac and an uninitialized variable.
Switchng mac tools to clang found these two:

1. Uninitialized variable in rpcin.c: clang's static analyzer seems
to have had trouble figuring out that it is impossible for execution
to reach the "if (status == FALSE) {" line without being set.
To appease it, initialize it to FALSE.

2. xdrproc_t: On many platforms, xdrproc_t is an unprototyped function
pointer that can take either 2 or 3 arguments. Apple changed their
implementation to always take 3 arguments and recommend replacing calls
with 2 arguments to add a 0 3rd argument.