Hgfs Server: add write op arg basic validate write op for security fixes
Add the basic write operation argument validation routine to obtain the data
sizes of the request and data to write for the supported HGFS protocol versions
(1 through 4). Add a call to it from the main write operation handler.

Note the argument validation checks will be added in the next change and moved
from the platform specific routines.