Thaw/resume guest file system when VMX aborts the operation.
When the operation is aborted by VMX after quiescing the
guest, we need to thaw/resume the guest.

In response to VMX message "vmbackup.abort", when guest sends
"vmbackup.eventSet req.aborted 4 Quiesce aborted." message,
it receives "Unknown command" response because VMX side RPC
handler for vmbackup event is removed before guest gets a chance
to respond. That is the behavior without this fix.
With this fix, in addition to the message above the new message
"vmbackup.eventSet req.done" added by this fix also receives
"Unknown command" response. We can fix these by keeping the RPC
handler for longer, but its not worth it because the operation
is being aborted anyway and vmbackup code in the guest is able
to tolerate these errors.