dictionary: Limit line data growth
The dictionary code keeps the raw input line around. Since the length
of the line is unknown at parsing time, a DynBuf is used to allocate
the memory for the line.

The DynBuf initial allocation is large - 1K - and this wastes space
when lines are rarely over 64 bytes long. The fix is to have the
dictionary code preallocate a 64 byte buffer before allowing appends
to the DynBuf. This will prevent many time consuming and memory
fragmenting realloc operations.