Remove Hostinfo_CPUCounts (hosted)
The Hostinfo_CPUCounts function has an unhappy history
where not all OSes (especially Mac) make it easy to
discover the CPU topology.

There is a good reason for this: the CPU topology generally
does not matter. In fact, it does not matter for us.
So let's remove this enumeration code and instead report
a single CPU package with nCores == nThreads == what the
host OS claims.  (Especially as the host OS does not really
expose cores versus threads information...). Then rip out
the underlying functions, all the way down.