Add xmlsec1 to open-vm-tools
Add the new files to open vm tools and add a switch to
turn it on.  xml-seccurity-c remains the default for
now.

Also clean up the service Makefile, which was using
the client lib (which also dragged in stuff it shouldn't have)
for common code instead of building it in.