VThreadBase signal counting: remove malloc
This is the easy first half of rewriting VThreadBase to not malloc
within signal handlers. The second half is still under construction.

The core problem is that malloc() inside a signal handler can
deadlock; consider especially the case of malloc() detecting
heap corruption -> SEGV -> signal handler -> malloc() and
self-deadlocking. Thus, the overall plan is to stop malloc()ing
and instead use TLS variables ... as the new comment points out,
we can now count on good TLS behavior since Linux and Windows
minimum OS versions bumped.

In the signal handler case this is easy: if __thread is available
we can just use that as a counter. Life is a little more tricky
without __thread (see: macOS and Android), where we need to use
pthread_{get,set}specific() calls. Use a constructor/destructor
to manage the TLS keys, which is a strategy dependent on the
TLS destructor being NULL. (Specifically, on library unload
the destructor function can remove the TLS key but cannot fire
the TLS destructor on all threads, which would leak).
One small note, the use of an atomic is not needed here because
only the current thread can set the value and, if interrupted by
a signal, the stack preserves the correct value.

Two other minor things deserve mention: first, remove the TID
parameter from VThreadBase_SetIsInSignal(); it was only used for
self. Second, re-ordered setting is-in-signal and the first
VThread_CurID() in the signal handler; this means the VThread_CurID()
implementation can check whether we are in a signal handler, a
feature that will matter in a future change.

Net result: the signal counting part can now run without any memory
allocations whatsoever. Memory allocation is thus needed only for
the thread ID and name. Stay tuned for TID/name in a future change.