Update VGAuth source for compilation with OpenSSL 1.0.x or 1.1.0
OpenSSL 1.1.0 is about to released.  With that release comes a number
of changes that include making several types opaque.  Along with this,
new or renamed setter/getter functions appear.   The API changes and
the impact on open-source products is discussed at:
https://wiki.openssl.org/index.php/1.1_API_Changes.  Updated man pages
are available at: https://www.openssl.org/docs/manmaster/

The impact to  open-vm-tools is that the message digest type MD_CTX
has become opague.  Local (stack) MD_CTX objects are not supported.
The required change is to dynamically allocate MD_CTX objects as needed
using the 1.1.0 API xxx_new() and xxx_free() or the pre-1.1.0 functions
xxx_create() and xxx_destroy().