Add string concatenation function to DynBuf
The DynBuf code operates on both strings and non-strings.  To add a
string at the end of a DynBuf is easy: DynBuf_AppendString() appends
a copy of the input string, including its NULL-termination.  Sometimes
strings are repeatedly concatenated, such that a version of this
function without NULL-termination is needed.

This change implements DynBuf_ConcatString() which explicitly does not
copy the NULL at the end of a string.