Update glib and pcre for FreeBSD and Solaris
Update glib, pcre, and additional libraries they require
for FreeBSD and Solaris.

With this change, a release note will be needed to advise
users to add the following lines to /etc/libmap.conf before
installing vmtools on FreeBSD 6.x:

 	[/usr/local/lib/vmware-tools/]
 	libpthread.so.2 libthr.so.2
 	libpthread.so   libthr.so

The presence of these lines will cause the run-time linker to use
libthr.so.2 whenever it encounter a dependency on libpthread.so.2
in an executable whose path starts with /usr/local/lib/vmware-tools.

The reason for the release note is as follows.  FreeBSD provides
a threads library called libthr.so.2.  In FreeBSD 6.x, this library
is separate from and incompatible with libpthread.so.2.  In FreeBSD
7 or later, libpthread.so.2 is a symlink to libthr.so.2.

vmtoolslib is linked against libthr.so.2, while libgthread-2.0.so.0
and libglib-2.0.so.0 are linked against libpthread.so.2.  On FreeBSD
7 or later, this works fine since libpthread.so.2 is a symlink to
libthr.so.2, as noted above.  However, on FreeBSD 6.x, vmtoolsd
core dumps because libthr.so.2 and libpthread.so.2 are separate
libraries, hence both are loaded and their incomatibility results
in a core dump.

Adding the above lines in /etc/libmap.conf will keep libpthread.so.2
from being loaded, thereby allowing vmtoolsd to run.