Properly configure xmlsec1 for CentOS
Some distros build xmlsec1 with dynamic crypto, so do things slightly
differently.

Also work around an issue with the way Redhat & SuSE package xmlsec1.
They built it with -DXMLSEC_NO_SIZE_T, but that flag isn't properly
exposed by xmlsec1-config.  This results in a bunch of xmlsec1 data
structures (anything with an xmlSecSize in it) being different between
the lib and our code, and it tends not to work when our code references
fields in the data structure.  This change always sets -DXMLSEC_NO_SIZE_T.