Hgfs FUSE client: fix broken attribute caching on writes
The FUSE write implementation incorrectly invalidated the cache of file
attributes. The return code for the write operation checked for only invalidating
attribute cache if the result was zero. However, this result is negative for
an error (in which case the cached attributes should not be invalidated) and anything
else is the number of bytes written.

The fix for the broken attribute caching on writes is made to check the result
for greater than zero so any success. If no bytes were actually written we take
the conservative approach of still invalidating the cache as so all the files
attributes for that file will be reset in the FUSE client. This should be rare
at best.