Hgfs Linux: Fix mounter versioning constants
Currently, to allow for the mount information structure, which is passed
from the user mode mounter application to the Linux and OSX HGFS kernel
mode drivers, to be modified and support backwards compatibility, the
structure is versioned.  However, the version field uses a badly named
constant implying it is the HGFS protocol version. This now clashes with
a new enum for the HGFS protocol version added for tracking and reporting
the protocol version in use.  The mount information structure version
should have the version constants named for their intended purpose
e.g. HGFS_MOUNTINFO_VERSION.

This change modifies the contant and uses an enum. This effects the mounter
and HGFS kernel and OSX kernel mode components.