Fix help messages related to xerces and xmlsecurity libraries.
When configuring open-vm-tools, if the system doesn't have
xmlsecurity and xerces libraries, the 'configure' step stops and
gives a hint that the user should try with '--without-xerces-c'
and '--without-xml-security-c' options. Those options are wrong
and the right ones are '--without-xmlsecurity' and '--without-xerces'.

Fixed the configure.ac script to print the proper '--without-xxxx'
messages in the warnings.