Tools Daemon: Do not use kernel client mount HGFS when FUSE unavailable
When FUSE is not installed in new OS versions that are kernel 4.0 and newer
the vmhgfs-fuse client is unavailable to use. Previously, the mount RPC handler
would fall back to mounting Shared Folders using the kernel client.
This really should not be done as it would entail the user running with an
untested vmhgfs kernel driver client or worse the driver fails to build.
Instead the RPC should really fail to mount.

The RPC Mount Shared Folders should only mount the Shared Folders using the
kernel client when FUSE is not supported by our HGFS FUSE client i.e., when
the Linux kernel version is older than 4.0.

To check for these conditions the vmhgfs-fuse component is run with the enabled
option which returns with success for enabled, 1 for not enabled (older OS kernel)
and 2 for newer kernel but FUSE is missing. These return values should be checked
and  only if value 1 is returned then the mount using the kernel client should be
attempted. Not as currently done with an unsuccessful result of 1 or 2.
This matches the installer which does the same check.

This fix is related to a bug which came about due to a user custom building
and installing a newer kernel 4.0+ on top of a 3.x kernel with tools, and without FUSE.
When the VM was rebooted using the 4.0+ kernel and running tools config script the
script will to try and use the HGFS kernel client.
The tools daemon RPC handler to mount the Shared Folders will also use the kernel
client, too, in this scenario.
This is bad as the HGFS kernel client is running untested by VMware.
Both issues need to be closed, this change addresses the tools daemon RPC handler only.