VThread: put thread names in thread-local storage
Next step in the "stop allocating per-thread memory" effort. This change
moves thread name storage to thread-local storage. (VThreadIDs are
the hardest and will come last). This is also only half the work; a
it adds new storage but does not remove the old storage, which was
left out to keep this change simpler.

For platforms that have thread-local storage, the implementation is
pretty straightforward, though initialization is inherently lazy.

For platforms lacking thread-local storage (at this point: iOS and
Android, see comment in vthreadBase.c), life is much harder.  Added
a very large comment contemplating different solutions and
ultimately concluding that unnamed threads on such a platform will
get a best-effort (read: racy) name from a global buffer. Note that
threads which are explicitly named WILL get a (non-racy) name, but
it requires an explicit naming attempt.