Move VThread_CurName() entirely to thread-local storage
A previous change switched VThreadBase to store thread names in
thread-local storage on most platforms (all except mobile). This
change is the follow-up: removing the old storage location for
thread names.

Briefly, this change removes the "name" field from VThreadBaseData and
handles all the fallout. That's mostly straightforward, though
lib/thread ended up storing a thread name in the "reserved" thread
structures - this is mostly so the name passed to VThread_ReserveID
would be preserved until VThread_CreateThread is called.

After this, vthreadBase.c complexity only remains for VThreadIDs, which
is the next step of this effort.