Allow enabling or disabling timesync multiple times.
Toolbox cmd sends a presumed old value for timesync GuestRpc
when enabling or disabling timesync. VMX side expects the guest
side to be aware of old value and therefore errors out whenever
guest repeats the operation.  Previously the implementation was
silently ignoring the error.

Avoid sending the guestRpc to enable or disable timesync when
it is already in the desired state (enabled or disabled respectively).
It is optimal and also avoids the unnecessary error.