Hgfs Server: Clean up logging macros and enable Hgfs Server tools logging
The HGFS server logging macros are scattered throughout and repeated
in each source file. This makes it difficult to modify between the
VMX implementation and the tools. The tools side suffers currently and
is thus hard to debug.

This is the first in a couple of small changes to clean up the logging.

First the cleanup by moving the server logging statements for LOG to a centralized
private header file for the server code and not repeated in every source file.

Second, move the HgfsServer LOG from hgfs to its own HgfsServer setting.
New VMX file setting is:
Loglevel.hgfsServer = "10"

The current existing setting of "hgfs" will capture logging for the VMX
such as the HGFS manager, policy and transport interfaces (VMCI/Backdoor).

Third, the HgfsServer LOG statements for the tools builds (which go into
both instances of the vmtoolsd sevice/daemon) and the upgrader application
are now mapped to g_debug and Debug function calls respectively.
This now allows the tools configuration logging to set the following which will capture all
the Hgfs server logging:
hgfsServer.level = "debug"

Follow up changes will be made to check the tools backdoor transport interface log setting
which currently uses something like:
hgfsd.level = "debug"