Specify the proper size while sending deploypkg.update.state RPC message.
Due to a recent change, the VMX handler for 'deployPkg.update.state' RPC
ignores the message if the specified argument size doesn't match with
the string size.

In the current code, in a certain error code path, deploypkg plugin
specifies one extra byte in the size for 'deploypkg.update.size' RPC
message.  This results in an error at the VMX layer.

Fixed the code to specify the proper string length while sending the RPC.