vm_{atomic,uninterruptible}.h: Flip the polarity of the 32-bit function names
Initially, we only had 32-bit atomic functions, so their names did not have a "32" suffix.

Then, we added 64-bit functions. So,
1) We used a "64" suffix for them.
2) We provided additional names with a "32" suffix for the 32-bit functions, for consistency.

The hope was that, over time, references to unsuffixed names would disappear.
This change sends a stronger signal to developpers that the suffixed names
are the ones to use, and that the unsuffixed names are the undesirable legacy
ones to avoid.