Add constants for macOS os major versions.
MacOS versions can get a bit confusing because there are two parallel
versioning schemes. There is the marketing version (e.g. 10.6, 10.7.5, etc.)
which is what everybody uses when discussing the OS but is very hard to get
at via code, and there is the kernel version (e.g. 10.0.0, 11.4.2) which
is what is returned by POSIX syscalls like uname. The confusion sets in
because in comments we tend to use the marketing version, but then the
code actually checks for the kernel version.

This change adds constants for the marketing version for comparisons with
HostInfo_OSVersion(0) (the major version), which makes checking for
10.6 vs 10.7 vs 10.8 etc. much more readable.