Add a generic interface for accepting backup manifest files.
The interface provided by VMX to allow VMTools to notify it that
there is a backup manifest file to be transferred from the guest
is specific to Windows guests.  A more generic interface is being
implemented to enable backup manifest files for Linux guests.  A
new vmbackup event is added for passing to vmx the guest's path
for the manifest file, along with code for handling the event.