Generate backup manifest files for quiesced snapshots on Linux.
After a successful quiesced snapshot of a Linux guest, the
snapshot is reported as not quiesced.  The underlying cause
is:

- a snapshot is identified as quiesced if and only if a backup
manifest file was provided for the snapshot; and

- backup manifest files are not provided for quiesced snapshots
on Linux.

To fix the problem, VMTools will generate backup manifest files
for quiesced snaphots taken on Linux, and VMX will accept and
store such files.  This change is for the VMTools side of the
change.  The VMX change defines a new VMBACKUP_EVENT value relied
on by the VMTools change.

Changes in this commit include:
- new source files syncManifest.h and syncManifest.c with code
to generate the manifest file, and updates to syncDriverOps.c
to invoke that code.
- a new "SyncDriver_GetAttr" interface in the syncDriver code,
used by the SyncManifest code to obtain the name of the backend
provider and whether or not it is capable of quiescing.
- makefile and configure boiler plate to add the two new source
files to open-vm-tools and tools-for-linux.
- a new tools.conf setting enableXmlManifest that specifies
whether VMTools generates backup manifests for Linux.  For now,
the default value of this setting will be false.