VMCI: Reset datagram buffer pointer on dealloc
A customer was seeing a guest crash when unloading the VMCI
driver - but only when upgrading to a specific version of UEK.
It turns out, that a BH may run during module unload,
and attempt to deliver datagrams to the datagram buffer after
it has been deallocated. It turns out that the tasklet may not
have been scheduled yet, so add a tasklet_disable to guard against
this. Also, reset the datagram buffer pointer when the buffer is
deallocated to indicate that it is no longer there.