Hgfs Server: Fix a memory leak of the transport connection
This change fixes a memory leak of the transport connection object
which was not being released on the final reference removal when
the HGFS server is being torn down.

This separates the initialization and destruction of the transport
object into their own transport init and exit functions.  Added the
lock destruct and free call of the transport object to the exit
function that were previously missing.  Corrected or added some
missing logging for debugging.

Additionally, fixed the VMCI interface to correct its tear down
order of functions which relied on the transport not being deleted.
This involved moving the draining of out of band HGFS server requests
or replies to the guest from being called after the HGFS server
transport has been destroyed.  Now it is moved to the correct place,
where the sequence is now:
- disconnect the HGFS server (i.e., stop generating any new request/replies out of band,
- drain the existing out of band HGFS requests/replies
- close the VMCI transport and shared memory callbacks
- close and teardown the HGFS server session and transport
- teardown the VMCI channel itself