optimize clamped uadd functions
For unsigned adds, if there's no overflow, the result will be larger
or equal than either source operand. Consequently, if there's an
overflow, the result will be smaller than both, and checking against
one source operand is sufficient. Found accidentally...
(With just one comparison, gcc is in fact able to figure out the
comparison is really the same as the add overflowing and will omit
the comparison. But it won't do this with two comparisons.)