Remove convertutf and bsd_printf
The license for convertutf.c and convertutf.h does not
explicitly permit, or prohibit, modification of the files,
and restricts use to "products supporting the Unicode
Standard". This is a problem for distribution with Debian,
which complies to strict guidelines. See details at
https://github.com/vmware/open-vm-tools/issues/148
and
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=823100

The code is not essential for open-vm-tools. It's needed by
code in bsd_printf.c and related code, which in turn is
needed by msgfmt.c and msgList.c. The latter code is not needed
for open-vm-tools, and can therefore be eliminated from
the open-vm-tools package.

When HAS_BSD_PRINTF isn't set, Str_Snprintf() and
Str_Vsnprintf will fall back to vsnprintf(), and
Str_Asprintf()/Str_Vasprintf() will fall back to
vasprintf(). Under Linux and FreeBSD these functions are
properly implemented, and support %ls and %S.