Hgfs Server: Remove the register share callback and usage
Remove the register share callback and usage which is an HGFS server
libray function passed back to the VMX Hgfs server manager and policy
manager. This tracked shares as they were individually added and removed.
It carried with it VMDB behavior which was erroneous for the HGFS server
share tracking.

This API has been replaced by piggybacking on the Hgfs server InvalidateObjects
callback which informs the Hgfs server when the VMDB has finished updating
the Hgfs policy manager with shares and the Policy manager's list is complete.
At this point the Hgfs server can close any open handles on stale shares and
remove those stale shares and add new shares.