Hgfs Server: Fix shares update for change notification
This change fixes UI updates of the shares which currently
update by a poorly designed algorithm which causes change
notification subscriptions to be deleted and not restored
and also unnecessary deactivation and activation of the
events generation.

Currently, the UI updates the HGFS shares in the HGFS server
via the RegisterShare callback. This is done from HGFS VMX
policy code which handles updates from the UI and maintains
the HGFS shares. The UI calls the policy code for each share
to remove it until the list of shares is empty. It then calls
back for each share to add the existing shares back to the list.
Currently, the policy code calls the HGFS server RegisterShare
callback to delete the share as it removes them and then again
to add each one back in that added to the policy shares list.
Unfortunately, as the shares are removed any existing subscribers
for change notifications on the deleted share are also destroyed.
Furthermore, when the shares list is empty, the count of subscribers
will also be zero at this point causing a deactivation of the
file change event generator. Once, the shares are added back in
via the RegisterShare callback no event subscribers will be
restored as they are not tracked and the event generation
will also not be activated due to that reason.

The fix is to make the shares tracking be handled in the same
manner as the HGFS server does for open handles on shares.
The HGFS policy waits until the shares list updates are complete
as the UI is done updating the shares. The HGFS policy then calls
the HGFS server InvalidateObjects callback with the new complete
list of shares which may or may not be different from the previous
list. The HGFS server InvalidateObjects goes through its current list
of shares and searches for each share in the new list of shares and
if it fails to locate it, then closes all open handles on that share.
Once the list has been exhausted, all open handles remaining will be
on shares that still exist.

The same callback is now used for resetting the list of shares for
the current subscribers of change events and not the RegisterShare
callback which should be removed altogether.  We now make sure that
the reset of file change event shares only removes shares not found
on the new list thereby only removing subscribers on those stale shares.
All subscribers for shares remaining on the lists are not deleted,
remain intact and unaffected.  Furthermore, file system event
generation will also not be deactivated unless all subscribers
happen to be for shares that were deleted.

This change is the first part which as code to piggyback on the HGFS
server InvalidateObjects callback to compare any shares in its list
with the new list and delete shares not found.  The RegisterSharecallback
function code is deleted and temporarily is an empty callback.
This will be deleted along with the callback code from the HGFS
policy manager in a subsequent change.

Details:
HgfsServerRegisterShare - is gutted and just an empty function just
   logging that it is called.
HgfsServerCleanupDeletedFolders is effectively renamed to
   HgfsServerSharesDeleteStale w/o the lock acquistion/release.
HgfsServerSharesDeleteStale now free the share name as well as the
   share object to stop memory leak.
HgfsServerShareAddInternal is extracted from the old defunct
   HgfsServerRegisterShare w/o the lock acquisition/release.
HgfsServerShareAdd acquires and releases the lock and calls
   HgfsServerShareAddInternal
HgfsServerSharesReset is created and called from the
   HgfsServerInvalidateObjects
HgfsServerEnumerateSharedFolders now calls the new HgfsServerShareAdd
   for all shares which is used to initialize the shares on HGFS server
   start.
HgfsServerGetLocalNameInfo now does not get the Shared Folder handle
   from the VMX policy host, instead getting it from the existing
   HgfsServerGetShareHandle (the server maintained list of shares).
   This is only used for setting subscribers from the client when
   using handles and not path names.
HgfsServerSetDirWatchByName now uses the handle returned by the
   HgfsServerGetLocalNameInfo call.
markedForDeletion is now removed as holding stale state is not needed
   since stale shares are deleted and resources freed immediately.